[TE] Support ScaleFabric SHCA NICs
Add USE_SHCA build option for ScaleFabric SHCA NICs, which use
extended 17-bit LIDs instead of standard 16-bit IB LIDs. When enabled,
convert LIDs via shca_17b_types.h helpers and widen lid fields to
uint32_t across Transfer Engine and TENT RDMA paths.

Gate yalantinglibs IBV RPC (YLT_ENABLE_IBV) behind USE_SHCA=OFF so SHCA
builds skip ib_socket-based RDMA RPC and fall back to TCP with a warning.

Adjust dependencies.sh to detect SHCA headers and skip conflicting system
RDMA packages (libibverbs-dev on Debian, rdma-core-devel on RHEL).

+++ mooncake-transfer-engine/tent/tests/rdma_transport_test.cpp
@@ -260,7 +260,7 @@ class RdmaContextTestPeer {
     }
 
     static void seedAddress(RdmaContext& context,
-                            const std::string& device_name, uint16_t lid,
+                            const std::string& device_name, uint32_t lid,
                             int gid_index, const ibv_gid& gid,
                             RdmaContext::DeviceStatus status) {
         std::lock_guard<std::mutex> guard(context.address_mutex_);
